[#noissue] Extract ApplicationMap domain
